requirements.txt
blinker==1.9.0
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
fpdf2==2.8.5
Jinja2==3.1.6
MarkupSafe==3.0.3
openpyxl==3.1.5
packaging==26.0
pillow==12.1.0
pluggy==1.6.0
pycparser==3.0
Pygments==2.19.2
PyMySQL==1.1.2
python-dotenv==1.1.1
pytz==2025.2
reportlab==4.4.9
SQLAlchemy==2.0.46
tomli==2.4.0
typing_extensions==4.15.0
Werkzeug==3.1.5
WTForms==3.2.1